coleta/requirements.txt
dash==2.0.0
dash-bootstrap-components==1.0.1
dash-bootstrap-templates==1.0.4
dash-core-components==2.0.0
dash-daq==0.5.0
pandas==0.24.2
plotly==5.3.1
pyzenhub==0.2.1
PyGithub==1.55
